RL now in RL class
